ref(explore): Lift toolbar visualize component
Lifting the visualize part of the toolbar into a shared folder. Logs will
prepare to use this component in the future.
