+++ static/app/views/explore/toolbar/index.tsx
@@ -1,10 +1,14 @@
 import styled from '@emotion/styled';
 
 import {defined} from 'sentry/utils';
+import {ToolbarVisualize} from 'sentry/views/explore/components/toolbar/toolbarVisualize';
+import {
+  useExploreVisualizes,
+  useSetExploreVisualizes,
+} from 'sentry/views/explore/contexts/pageParamsContext';
 import {ToolbarGroupBy} from 'sentry/views/explore/toolbar/toolbarGroupBy';
 import {ToolbarSaveAs} from 'sentry/views/explore/toolbar/toolbarSaveAs';
 import {ToolbarSortBy} from 'sentry/views/explore/toolbar/toolbarSortBy';
-import {ToolbarVisualize} from 'sentry/views/explore/toolbar/toolbarVisualize';
 
 type Extras = 'equations';
 
@@ -14,9 +18,16 @@ interface ExploreToolbarProps {
 }
 
 export function ExploreToolbar({extras, width}: ExploreToolbarProps) {
+  const visualizes = useExploreVisualizes();
+  const setVisualizes = useSetExploreVisualizes();
+
   return (
     <Container width={width}>
-      <ToolbarVisualize allowEquations={extras?.includes('equations') || false} />
+      <ToolbarVisualize
+        visualizes={visualizes}
+        setVisualizes={setVisualizes}
+        allowEquations={extras?.includes('equations') || false}
+      />
       <ToolbarGroupBy autoSwitchToAggregates />
       <ToolbarSortBy />
       <ToolbarSaveAs />
